refactor(challenge): ParsedCommentBox의 parseLink 순수함수를 utils로 분리 (구조정리 P2)
컴포넌트 파일에 섞여 있던 parseLink(순수 텍스트→링크 변환)를
my-challenge/utils/parseLink.tsx로 분리. ParsedCommentBox는 컴포넌트만 유지.
동작 변경 없음.
